requirements.txt
playwright>=1.42.0
google-generativeai>=0.4.1
pandas>=2.2.2 
openpyxl>=3.1.2
streamlit>=1.32.2